Update locators and split actions into separate files for better organization.
